Separate core SDK and Merchant-Backend-specific utilities to their own libraries
