+++ Podfile
@@ -1,25 +1,7 @@
-# Uncomment the next line to define a global platform for your project
-# platform :ios, '9.0'
-
-target 'SwedbankPaySDK' do
-  # Comment the next line if you don't want to use dynamic frameworks
+abstract_target 'NeedsAlamofire' do
   use_frameworks!
-
-  # Pods for SwedbankPaySDK
-  pod 'Alamofire', '~> 5.2'
-
-  target 'SwedbankPaySDKTests' do
-    inherit! :search_paths
-    # Pods for testing
-    
-    pod 'Alamofire', '~> 5.2'
-  end
-  
-  target 'SwedbankPaySDKUITestHost' do
-    inherit! :search_paths
-    # Pods for UI testing
-    
-    pod 'Alamofire', '~> 5.2'
-  end
-
+  pod 'Alamofire', '~> 5.4'
+  target 'SwedbankPaySDKMerchantBackend'
+  target 'SwedbankPaySDKTests'
+  target 'SwedbankPaySDKUITestHost'
 end
