+++ SwedbankPaySDK.podspec
@@ -22,9 +22,9 @@ The Swedbank Pay iOS SDK enables simple embedding of Swedbank Pay Checkout to an
 
   s.ios.deployment_target = '10.0'
   s.swift_versions = '5.0', '5.1'
+  
   s.source_files = 'SwedbankPaySDK/Classes/**/*'
   s.resources = 'SwedbankPaySDK/Resources/**/*'
   
   s.frameworks = 'UIKit', 'WebKit'
-  s.dependency 'Alamofire', '~> 5.2'
 end
