+++ SwedbankPaySDK/Classes/SwedbankPaySDK+Extensions/PaymentOrder.swift
@@ -603,82 +603,6 @@ public extension SwedbankPaySDK {
     }
 }
 
-public extension SwedbankPaySDK.PaymentOrderUrls {
-    private static func buildCompleteUrl(configuration: SwedbankPaySDK.MerchantBackendConfiguration) -> URL {
-        return URL(string: "complete", relativeTo: configuration.backendUrl)!
-    }
-    private static func buildCancelUrl(configuration: SwedbankPaySDK.MerchantBackendConfiguration) -> URL {
-        return URL(string: "cancel", relativeTo: configuration.backendUrl)!
-    }
-    private static func buildPaymentUrl(configuration: SwedbankPaySDK.MerchantBackendConfiguration, language: SwedbankPaySDK.Language, id: String) -> URL {
-        var components = URLComponents()
-        components.path = "sdk-callback/ios-universal-link"
-        var queryItems: [URLQueryItem] = [
-            .init(name: "scheme", value: configuration.callbackScheme),
-            .init(name: "language", value: language.rawValue),
-            .init(name: "id", value: id)
-        ]
-        if let appName = getAppName() {
-            queryItems.append(.init(name: "app", value: appName))
-        }
-        components.queryItems = queryItems
-        return components.url(relativeTo: configuration.backendUrl)!
-    }
-    private static func getAppName() -> String? {
-        let bundle = Bundle.main
-        let displayName = bundle.object(forInfoDictionaryKey: "CFBundleDisplayName") as? String
-        return displayName ?? bundle.object(forInfoDictionaryKey: "CFBundleName") as? String
-    }
-    
-    /// Convenience initializer that generates a set of urls
-    /// for a payment using `MerchantBackendConfiguration`
-    ///  - parameter configuration: the MerchantBackendConfiguration where this payment is to be used
-    ///  - parameter language: the language of the payment
-    ///  - parameter callbackUrl: the callbackUrl to set for the payment
-    ///  - parameter termsOfServiceUrl: the Terms of Service url of the payment
-    ///  - parameter identifier: an unique identifier that is used to identify this payment **inside this application**
-    init(
-        configuration: SwedbankPaySDK.MerchantBackendConfiguration,
-        language: SwedbankPaySDK.Language,
-        callbackUrl: URL? = nil,
-        termsOfServiceUrl: URL? = nil,
-        identifier: String = UUID().uuidString
-    ) {
-        self.init(
-            configuration: configuration,
-            language: language,
-            hostUrl: configuration.backendUrl,
-            callbackUrl: callbackUrl,
-            termsOfServiceUrl: termsOfServiceUrl
-        )
-    }
-    
-    /// Convenience initializer that generates a set of urls
-    /// for a payment using `MerchantBackendConfiguration`
-    ///  - parameter configuration: the MerchantBackendConfiguration where this payment is to be used
-    ///  - parameter language: the language of the payment
-    ///  - parameter hostUrl: the url to set in the hostUrls of the payment.
-    ///   This will also become the `webViewBaseURL` of the `ViewPaymentOrderInfo` created for this payment
-    ///  - parameter callbackUrl: the callbackUrl to set for the payment
-    ///  - parameter termsOfServiceUrl: the Terms of Service url of the payment
-    ///  - parameter identifier: an unique identifier that is used to identify this payment **inside this application**
-    init(
-        configuration: SwedbankPaySDK.MerchantBackendConfiguration,
-        language: SwedbankPaySDK.Language,
-        hostUrl: URL,
-        callbackUrl: URL? = nil,
-        termsOfServiceUrl: URL? = nil,
-        identifier: String = UUID().uuidString
-    ) {
-        self.hostUrls = [hostUrl]
-        self.completeUrl = SwedbankPaySDK.PaymentOrderUrls.buildCompleteUrl(configuration: configuration)
-        self.cancelUrl = SwedbankPaySDK.PaymentOrderUrls.buildCancelUrl(configuration: configuration)
-        self.paymentUrl = SwedbankPaySDK.PaymentOrderUrls.buildPaymentUrl(configuration: configuration, language: language, id: identifier)
-        self.callbackUrl = callbackUrl
-        self.termsOfServiceUrl = termsOfServiceUrl
-    }
-}
-
 public extension SwedbankPaySDK.ShipIndicator {
     enum Raw : String, Codable {
         case ShipToBillingAddress = "01"
