+++ SwedbankPaySDKMerchantBackend/Classes/MerchantBackendApi.swift
@@ -14,6 +14,7 @@
 // limitations under the License.
 
 import Foundation
+import SwedbankPaySDK
 import Alamofire
 
 struct MerchantBackendApi {
