+++ SwedbankPaySDKMerchantBackend/Classes/MerchantBackendError.swift
@@ -14,6 +14,7 @@
 // limitations under the License.
 
 import Foundation
+import SwedbankPaySDK
 
 public extension SwedbankPaySDK {
     /// Ways that the SwedbankPaySDK.MerchantBackendConfiguration
