+++ SwedbankPaySDKMerchantBackend/Classes/Models/PayerOwnedPaymentTokensResponse.swift
@@ -14,7 +14,16 @@
 // limitations under the License.
 
 import Foundation
+import SwedbankPaySDK
 
-struct ConsumerSession: Decodable {
-    var operations: [SwedbankPaySDK.Operation]
+public extension SwedbankPaySDK {
+    /// Response to MerchantBackend.getPayerOwnedPaymentTokens
+    struct PayerOwnedPaymentTokensResponse: Decodable {
+        /// The response payload.
+        public var payerOwnedPaymentTokens: PayerOwnedPaymentTokens
+        
+        /// Operations you can perform on the whole list of tokens.
+        /// Note that you generally cannot call these from your mobile app.
+        public var operations: [SwedbankPaySDK.Operation]?
+    }
 }
