+++ SwedbankPaySDKMerchantBackend/Classes/Models/SetInstrumentLink.swift
@@ -1,4 +1,3 @@
-//
 // Copyright 2020 Swedbank AB
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,23 +13,33 @@
 // limitations under the License.
 
 import Foundation
+import SwedbankPaySDK
 
-struct PaymentOrdersLink: Link {
+struct SetInstrumentLink: Link {
     let href: URL
     
-    func post(
+    /// - returns: function that cancels this operation
+    func patch(
         api: MerchantBackendApi,
-        paymentOrder: SwedbankPaySDK.PaymentOrder,
+        instrument: SwedbankPaySDK.Instrument,
         userData: Any?,
         completion: @escaping (Result<PaymentOrderIn, SwedbankPaySDK.MerchantBackendError>) -> Void
-    ) {
-        let body = Body(paymentorder: paymentOrder)
-        _ = request(api: api, method: .post, body: body, completion: completion) { decorator, request in
-            decorator.decorateCreatePaymentOrder(request: &request, paymentOrder: paymentOrder, userData: userData)
+    ) -> SwedbankPaySDKRequest? {
+        let body = Body(instrument: instrument)
+        return request(api: api, method: .patch, body: body, completion: completion) { decorator, request in
+            decorator.decoratePaymentOrderSetInstrument(request: &request, instrument: instrument, userData: userData)
         }
     }
     
     private struct Body: Encodable {
-        let paymentorder: SwedbankPaySDK.PaymentOrder
+        init(instrument: SwedbankPaySDK.Instrument) {
+            paymentorder = PaymentOrder(instrument: instrument)
+        }
+        
+        private let paymentorder: PaymentOrder
+        private struct PaymentOrder: Encodable {
+            let operation = "SetInstrument"
+            let instrument: SwedbankPaySDK.Instrument
+        }
     }
 }
